1.0.0 - Reintroduced separate folder for unit tests
