+++ src/main/handlers/cookieHandler.js
