+++ src/main/handlers/manifestHandler.js
