+++ src/main/handlers/userPreferencesHandler.js
