+++ src/test/common.test.js
@@ -15,3 +15,4 @@ export const deleteAllCookies = () => {
         document.cookie = name + "=;expires=Thu, 01 Jan 1970 00:00:00 GMT";
     }
 }
+
