+++ src/test/models/config.test.js
@@ -1,4 +1,4 @@
-import Config from "./config";
+import Config from "../../main/models/config";
 
 describe('Config', () => {
